changed the values to double

+++ src/main/java/com/co2/models/AbstractCo2Measurments.java
@@ -1,26 +1,26 @@
 package com.co2.models;
 
 public abstract class AbstractCo2Measurments {
-    private Long maxCO2Units;
-    private Long minCO2Units;
-    private Long avgCO2Units;
+    private Double maxCO2Units;
+    private Double minCO2Units;
+    private Double avgCO2Units;
 	
-	public Long getMaxCO2Units() {
+	public Double getMaxCO2Units() {
 		return maxCO2Units;
 	}
-	public void setMaxCO2Units(Long maxCO2Units) {
+	public void setMaxCO2Units(Double maxCO2Units) {
 		this.maxCO2Units = maxCO2Units;
 	}
-	public Long getMinCO2Units() {
+	public Double getMinCO2Units() {
 		return minCO2Units;
 	}
-	public void setMinCO2Units(Long minCO2Units) {
+	public void setMinCO2Units(Double minCO2Units) {
 		this.minCO2Units = minCO2Units;
 	}
-	public Long getAvgCO2Units() {
+	public Double getAvgCO2Units() {
 		return avgCO2Units;
 	}
-	public void setAvgCO2Units(Long avgCO2Units) {
+	public void setAvgCO2Units(Double avgCO2Units) {
 		this.avgCO2Units = avgCO2Units;
 	}
